Move plugin package store into plugins service
Move provider package install/extract helpers out of gestaltd/internal/pluginstore and into services/plugins/store.

Interface diff:

- import "github.com/valon-technologies/gestalt/server/internal/pluginstore"

- pluginstore.Install(packagePath, destDir)

- *pluginstore.InstalledPlugin

+ import "github.com/valon-technologies/gestalt/server/services/plugins/store"

+ store.Install(packagePath, destDir)

+ *store.InstalledPlugin

Example:

installed, err := store.Install(download.LocalPath, destDir)
