+++ gestaltd/services/plugins/store/store.go
@@ -1,4 +1,4 @@
-package pluginstore
+package store
 
 import (
 	"crypto/sha256"
